+++ packages/adhere/script/build.js
@@ -58,6 +58,7 @@ const namedMap = new Map([
   ['@baifendian/adhere-ui-swipeout', 'SwipeOut'],
   ['@baifendian/adhere-ui-polygonselection', 'PolygonSelection'],
   ['@baifendian/adhere-ui-playground', 'PlayGround'],
+  ['@baifendian/adhere-ui-tablegridlayout', 'TableGridLayout'],
   // ['@baifendian/adhere-ui-bmap', 'BMap'],
   ['@baifendian/adhere-ui-forceupdate', 'ForceUpdate'],
   ['@baifendian/adhere-util', 'Util'],