+++ packages/adhere/src/table-grid-layout/index.js
@@ -1,0 +1,2 @@
+import Model from '@baifendian/adhere-ui-tablegridlayout/es';
+export default Model;